Changes after review - Moved some resources. - Changed xml documentation. - Improved tests. Related to Issue [WTI-616]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@3718 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 4ceda0de586616bdfe6a5c119790dd7d61c1ea3c
Former-commit-id: 2342043eca8ee4a73376853ad1232fee814eb51c

+++ Ringtoets/HydraRing/src/Ringtoets.HydraRing.Calculation/Data/Input/ExceedanceProbabilityCalculationInput.cs
@@ -31,7 +31,7 @@ namespace Ringtoets.HydraRing.Calculation.Data.Input
         /// Creates a new instance of the <see cref="ExceedanceProbabilityCalculationInput"/> class.
         /// </summary>
         /// <param name="hydraulicBoundaryLocationId">The id of the hydraulic station to use during the calculation.</param>
-        protected ExceedanceProbabilityCalculationInput(int hydraulicBoundaryLocationId) : base(hydraulicBoundaryLocationId) {}
+        protected ExceedanceProbabilityCalculationInput(long hydraulicBoundaryLocationId) : base(hydraulicBoundaryLocationId) {}
 
         public override int CalculationTypeId
         {
